requirements.txt
pyyaml
streamlit
openai==0.28.1
